move operator structs to pred and ref structs
